+++ app/src/main/java/de/baumann/browser/view/AdapterTabs.java
@@ -9,8 +9,6 @@ import android.view.View;
 import android.widget.ImageView;
 import android.widget.TextView;
 
-import androidx.core.content.ContextCompat;
-
 import com.google.android.material.card.MaterialCardView;
 
 import de.baumann.browser.R;
@@ -27,8 +25,6 @@ class AdapterTabs {
     private TextView albumUrl;
     private BrowserController browserController;
     private MaterialCardView albumCardView;
-    private ImageView albumClose;
-    private ImageView albumFavicon;
 
     AdapterTabs(Context context, AlbumController albumController, BrowserController browserController) {
         this.context = context;
@@ -58,8 +54,7 @@ class AdapterTabs {
         albumUrl = albumView.findViewById(R.id.dateView);
         albumCardView = albumView.findViewById(R.id.albumCardView);
 
-        albumFavicon = albumView.findViewById(R.id.faviconView);
-        albumClose = albumView.findViewById(R.id.iconView);
+        ImageView albumClose = albumView.findViewById(R.id.iconView);
         albumClose.setImageResource(R.drawable.icon_tab_remove);
         albumClose.setVisibility(View.VISIBLE);
         albumClose.setOnClickListener(view -> {
@@ -70,19 +65,11 @@ class AdapterTabs {
 
     public void activate() {
         TypedValue typedValue = new TypedValue();
-        context.getTheme().resolveAttribute(R.attr.colorSecondary, typedValue, true);
+        context.getTheme().resolveAttribute(R.attr.colorSecondaryContainer, typedValue, true);
         int color = typedValue.data;
         context.getTheme().resolveAttribute(R.attr.colorSurface, typedValue, true);
-        int color2 = typedValue.data;
-        if (albumFavicon.getDrawable() == ContextCompat.getDrawable(context, R.drawable.icon_image_broken)) {
-            albumFavicon.setColorFilter(color2, android.graphics.PorterDuff.Mode.SRC_IN);
-        }
-
         albumCardView.setCardBackgroundColor(color);
-        albumClose.setColorFilter(color2, android.graphics.PorterDuff.Mode.SRC_IN);
         albumTitle.setTypeface(null, Typeface.BOLD);
-        albumTitle.setTextColor(color2);
-        albumUrl.setTextColor(color2);
         albumView.setOnClickListener(view -> {
             albumCardView.setCardBackgroundColor(color);
             browserController.hideOverview();
@@ -91,19 +78,11 @@ class AdapterTabs {
 
     void deactivate() {
         TypedValue typedValue = new TypedValue();
-        context.getTheme().resolveAttribute(R.attr.colorSecondaryContainer, typedValue, true);
+        context.getTheme().resolveAttribute(R.attr.colorSurfaceDim, typedValue, true);
         int color = typedValue.data;
         context.getTheme().resolveAttribute(R.attr.colorOnSurfaceVariant, typedValue, true);
-        int color2 = typedValue.data;
-        if (albumFavicon.getDrawable() == ContextCompat.getDrawable(context, R.drawable.icon_image_broken)) {
-            albumFavicon.setColorFilter(color2, android.graphics.PorterDuff.Mode.SRC_IN);
-        }
-
         albumCardView.setCardBackgroundColor(color);
-        albumClose.setColorFilter(color2, android.graphics.PorterDuff.Mode.SRC_IN);
         albumTitle.setTypeface(null, Typeface.NORMAL);
-        albumTitle.setTextColor(color2);
-        albumUrl.setTextColor(color2);
         albumView.setOnClickListener(view -> {
             browserController.showAlbum(albumController);
             browserController.hideOverview();